refactor: export helper methods as part of a named object
