--- tests/utils/index.ts
@@ -1,75 +1,0 @@
-import { Page } from '@playwright/test';
-import { FakeDataFactory } from '../../factories/fakeDataFactory';
-import { AdminHomePage } from '../../pageObjectModels/adminHomePage';
-import { AppAdminPage } from '../../pageObjectModels/appAdminPage';
-import { AppsAdminPage } from '../../pageObjectModels/appsAdminPage';
-import { DashboardPage } from '../../pageObjectModels/dashboardPage';
-
-/**
- *
- */
-export async function navigateToAdminHomePage(sysAdminPage: Page) {
-  const dashboardPage = new DashboardPage(sysAdminPage);
-  await dashboardPage.goto();
-  await dashboardPage.page.waitForLoadState();
-  await dashboardPage.sharedNavPage.adminGearIcon.click();
-}
-
-/**
- * Helper function to create an app, add app to delete list, and return objects needed for tests.
- */
-export async function createAppForTest(
-  sysAdminPage: Page,
-  appsToDelete: string[]
-) {
-  const adminHomePage = new AdminHomePage(sysAdminPage);
-  const appAdminPage = new AppAdminPage(sysAdminPage);
-  const appName = FakeDataFactory.createFakeAppName();
-  appsToDelete.push(appName);
-
-  await adminHomePage.page.waitForLoadState();
-  await adminHomePage.createAppUsingHeaderCreateButton(appName);
-
-  return {
-    adminHomePage,
-    appAdminPage,
-    appName,
-  };
-}
-
-/**
- * Helper function to delete apps created during tests.
- */
-export async function deleteAppsForTest(
-  sysAdminPage: Page,
-  appsToDelete: string[]
-) {
-  const appsAdminPage = new AppsAdminPage(sysAdminPage);
-  await appsAdminPage.goto();
-  await appsAdminPage.page.waitForLoadState();
-
-  for (const appName of appsToDelete) {
-    const appRow = appsAdminPage.appGrid
-      .getByRole('row', { name: appName })
-      .first();
-    const appDeleteButton = appRow.getByTitle('Delete App');
-    await appsAdminPage.page.waitForLoadState('networkidle');
-
-    // eslint-disable-next-line playwright/no-force-option
-    await appRow.hover({ force: true });
-    // eslint-disable-next-line playwright/no-force-option
-    await appDeleteButton.click({ force: true });
-
-    await appsAdminPage.deleteAppDialog.confirmationInput.pressSequentially(
-      'OK',
-      {
-        delay: 100,
-      }
-    );
-
-    // eslint-disable-next-line playwright/no-force-option
-    await appsAdminPage.deleteAppDialog.deleteButton.click({ force: true });
-    await appsAdminPage.deleteAppDialog.waitForModalToBeDismissed();
-    await appsAdminPage.page.waitForLoadState('networkidle');
-  }
-}
